+++ src/main/java/com/extentech/toolkit/FastAddVector.java
@@ -61,13 +61,13 @@ public class FastAddVector extends SpecialArrayList implements java.io.Serializa
 	 */
 	public boolean addOrderedDouble( Double obj )
 	{
-		double d = obj.doubleValue();
+		double d = obj;
 		try
 		{
 			for( int i = 0; i < super.size(); i++ )
 			{
 				Double dd = (Double) super.get( i );
-				if( dd.doubleValue() > d )
+				if( dd > d )
 				{
 					super.add( i, obj );
 					return true;
